backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# AI/ML - Updated to support Python 3.12
tensorflow>=2.16.0
numpy>=1.26.0
pillow>=10.0.0

# Database (optional)
pymongo==4.6.0
dnspython>=2.6.0

# Utilities
python-dotenv==1.0.0
